+++ doc/release-notes.md
@@ -4,4 +4,6 @@ Bitcoin ABC version 0.30.6 is now available from:
 
   <https://download.bitcoinabc.org/0.30.6/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features and fixes:
+ - A new `getinfo` RPC has been added to retrieve basic information about the
+   node.